User registration

+++ server/controllers/NoteController.js
@@ -1,8 +1,12 @@
 const Note = require('../models/note')
+const jwt = require('jsonwebtoken');
 
 createNote = (req, res) => {
     const body = req.body
-    console.log('create note' + req.body);
+   // 'create note' + req.body
+    if (!verifyToken(req, res)) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
     if (!body) {
         return res.status(400).json({
             success: false,
@@ -12,7 +16,7 @@ createNote = (req, res) => {
 
     const note = new Note(body);
 
-    if (!note) {
+    if (!note || !note.userId) {
         return res.status(400).json({ success: false, error: err })
     }
 
@@ -36,9 +40,11 @@ createNote = (req, res) => {
 
 updateNote = async (req, res) => {
     const body = req.body;
-    console.log('update note' + req.params.id);
-
-    console.log(req.body);
+   //'update note' + req.params.id
+    if (!verifyToken(req, res)) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
+   //req.body
     if (!body) {
         return res.status(400).json({
             success: false,
@@ -76,7 +82,10 @@ updateNote = async (req, res) => {
 };
 
 deleteNote = async (req, res) => {
-    console.log('delete Note' + req.params.id);
+    //'delete Note' + req.params.id
+    if (!verifyToken(req, res)) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
     await Note.findOneAndDelete({ _id: req.params.id }, (err, note) => {
         if (err) {
             console.log('error during note deleting' + err);
@@ -87,8 +96,10 @@ deleteNote = async (req, res) => {
 };
 
 getNoteById = async (req, res) => {
-    console.log('get Note by id: ' + req.params.id);
-
+   // 'get Note by id: ' + req.params.id
+    if (!verifyToken(req, res)) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
     await Note.findOne({ _id: req.params.id }, (err, note) => {
         if (err) {
             console.log('error during get Note by id: ' + err);
@@ -100,8 +111,12 @@ getNoteById = async (req, res) => {
 }
 
 getNoteList = async (req, res) => {
-    console.log('get Note list: ');
-    await Note.find({}, (err, notes) => {
+    //'get Note list: '
+   // req.params.userId
+    if (!verifyToken(req, res) || !req.params.userId) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
+    await Note.find({userId: req.params.userId}, (err, notes) => {
         if (err) {
             console.log('error during get Note list: ' + err);
             return res.status(400).json({ success: false, error: err })
@@ -111,8 +126,11 @@ getNoteList = async (req, res) => {
 }
 
 deleteClearNotes = async (req, res) => {
-    console.log('delete completed Notes');
-    // console.log('delete completed Notes' + req.params);
+   // 'delete completed Notes'
+    // 'delete completed Notes' + req.params
+    if (!verifyToken(req, res)) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
     await Note.deleteMany({
         completed: true
     }, (err, note) => {
@@ -132,10 +150,12 @@ deleteClearNotes = async (req, res) => {
 };
 
 getFilter =  async (req, res) => {
-    console.log('get filter ',req.params.completed );
-
-        console.log('get Note list: ');
-        await Note.find({completed: req.params.completed}, (err, notes) => {
+   // 'get filter ',req.params.completed, req.params.userId
+    if (!verifyToken(req, res) || !req.params.userId) {
+        return res.json({status: 403, success: false, message: 'Failed to authenticate token.'});
+    }
+        //'get Note list: '
+        await Note.find({completed: req.params.completed, userId: req.params.userId}, (err, notes) => {
             if (err) {
                 console.log('error during get filter: ' + err);
                 return res.status(400).json({ success: false, error: err })
@@ -145,8 +165,24 @@ getFilter =  async (req, res) => {
 
 };
 
+verifyToken = (req, res) => {
+    const token = req.body.token || req.query.token || req.headers['x-access-token'] || req.headers.token;
+   // "router use", token
+    if (token) {
+        jwt.verify(token, 'superSecret', function (err, decoded) {
+            if (err) {
+                return false;
+            } else {
+                req.decoded = decoded;
+                console.log(decoded);
+            }
+        });
+    } else {
 
-
+        return false;
+    }
+    return true;
+};
 
 module.exports = {
     createNote,
